Move named groups test setup into setup.py
