+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/graphql_context_test_suite.py
@@ -857,22 +857,10 @@ ExecutingGraphQLContextTestMatrix = make_graphql_context_test_suite(
 )
 
 
-def get_loadable_target_for_file(filename: str) -> LoadableTargetOrigin:
-    return LoadableTargetOrigin(
-        executable_path=sys.executable,
-        python_file=file_relative_path(__file__, filename),
-    )
-
-
-CrossRepoDepsGraphQLContextTestMatrix = make_graphql_context_test_suite(
-    context_variants=GraphQLContextVariant.all_executing_variants(
-        target=get_loadable_target_for_file("cross_repo_asset_deps.py")
-    )
+all_repos_loadable_target = LoadableTargetOrigin(
+    executable_path=sys.executable,
+    python_file=file_relative_path(__file__, "cross_repo_asset_deps.py"),
 )
-
-# Repository for testing named asset groups
-NamedAssetGroupsGraphQLContextTestMatrix = make_graphql_context_test_suite(
-    context_variants=GraphQLContextVariant.all_executing_variants(
-        target=get_loadable_target_for_file("named_asset_groups_repo.py")
-    )
+AllRepositoryGraphQLContextTestMatrix = make_graphql_context_test_suite(
+    context_variants=GraphQLContextVariant.all_executing_variants(target=all_repos_loadable_target)
 )
