+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/setup.py
@@ -1632,6 +1632,33 @@ def unconnected(context):
 asset_group_job = AssetGroup([foo, bar, foo_bar, baz, unconnected]).build_job("foo_job")
 
 
+@asset(group_name="group_1")
+def grouped_asset_1():
+    return 1
+
+
+@asset(group_name="group_1")
+def grouped_asset_2():
+    return 1
+
+
+@asset
+def ungrouped_asset_3():
+    return 1
+
+
+@asset(group_name="group_2")
+def grouped_asset_4():
+    return 1
+
+
+# For now the only way to add assets to repositories is via AssetGroup
+# When AssetGroup is removed, these assets should be added directly to repository_with_named_groups
+named_groups_job = AssetGroup(
+    [grouped_asset_1, grouped_asset_2, ungrouped_asset_3, grouped_asset_4]
+).build_job("named_groups_job")
+
+
 @repository
 def empty_repo():
     return []
@@ -1694,6 +1721,7 @@ def define_pipelines():
         failure_assets_job,
         asset_group_job,
         hanging_graph_asset_job,
+        named_groups_job,
     ]
